ingest/requirements.txt
podaac-data-subscriber
earthaccess
google-cloud-pubsub